# Conflicts:
#	CMakeLists.txt
#	include/clients/common/Result.hpp
#	include/clients/nrt/BufferComposeNRT.hpp
#	include/clients/nrt/BufferStats.hpp
#	include/clients/nrt/FluidNRTClientWrapper.hpp
#	include/clients/nrt/NMFClient.hpp
#	include/clients/nrt/NoveltyClient.hpp
#	include/clients/rt/AmpSlice.hpp
#	include/clients/rt/BaseSTFTClient.hpp
#	include/clients/rt/BufferedProcess.hpp
#	include/clients/rt/GainClient.hpp
#	include/clients/rt/HPSSClient.hpp
#	include/clients/rt/LoudnessClient.hpp
#	include/clients/rt/MFCCClient.hpp
#	include/clients/rt/MelBandsClient.hpp
#	include/clients/rt/NMFFilter.hpp
#	include/clients/rt/NMFMatch.hpp
#	include/clients/rt/OnsetSlice.hpp
#	include/clients/rt/PitchClient.hpp
#	include/clients/rt/SinesClient.hpp
#	include/clients/rt/SpectralShapeClient.hpp
#	include/clients/rt/TransientClient.hpp
#	include/clients/rt/TransientSlice.hpp

+++ include/clients/common/ParameterTrackChanges.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cmath>
 #include <tuple>
 #include <utility>
 
@@ -36,8 +37,8 @@ private:
     bool allSame = true;
     using std::get;
     std::initializer_list<int>{(allSame = allSame && (get<Is>(mValues) == args), 0)...};
-    std::initializer_list<int>{(get<Is>(mSignums) = std::copysign(1, args - get<Is>(mValues)), 0)...};
-    std::initializer_list<int>{(get<Is>(mValues) = args, 0)...};
+    std::initializer_list<int>{(get<Is>(mSignums) = static_cast<int>(std::copysign(1, args - get<Is>(mValues))), 0)...};
+    std::initializer_list<int>{(get<Is>(mValues)  = args, 0)...};
     return !allSame;
   }
